+++ flutter-hms-adsprime/example/lib/pages/vast_ad_page.dart
@@ -52,12 +52,12 @@ class _VastAdPageBodyState extends State<VastAdPageBody>
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance?.addObserver(this);
+    WidgetsBinding.instance.addObserver(this);
   }
 
   @override
   void dispose() {
-    WidgetsBinding.instance?.removeObserver(this);
+    WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }