+++ flutter-hms-adsprime/lib/src/adslite/instream/instream_ad_view.dart
@@ -193,4 +193,17 @@ class InstreamAdViewController {
       'unmute',
     );
   }
+
+  Future<void> showTransparencyDialog({List<int>? location}) async {
+    return await _channel.invokeMethod(
+      'showTransparencyDialog',
+      location,
+    );
+  }
+
+  Future<void> hideTransparencyDialog() async {
+    return await _channel.invokeMethod(
+      'hideTransparencyDialog',
+    );
+  }
 }